move ispam func to utils
